careyou/rag/requirements.txt
loguru
langchain==0.1.13
langchain_community==0.0.29
langchain_core==0.1.33
langchain_openai==0.0.8
langchain_text_splitters==0.0.1
FlagEmbedding==1.2.8
unstructured==0.12.6
PyJWT
faiss-gpu  # faiss-cpu for device without gpu
